% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.round}
\alias{is.round}
\title{Checks if information is lost by rounding}
\usage{
is.round(x, digits = 0)
}
\arguments{
\item{x}{A vector.}

\item{digits}{How many digits to round to.}
}
\description{
This function takes a vector and checks if any information is lost by rounding to a certain number of digits.
}
\details{
Returns \code{TRUE} if rounding to \code{digits} digits after the decimal can be done without losing information.
}
\examples{
is.round(1:5)

x <- c(1, 1.2, 1.23)
is.round(x)
is.round(x,digits=2)
}
