% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pctile}
\alias{pctile}
\title{Returns a vector of 100 percentiles}
\usage{
pctile(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
This function calculates 100 percentiles of a vector and returns all of them.
}
\details{
This function just shorthand for \code{quantile(x,1:100/100)}, with a shorter name for reference in the \code{vtable} or \code{sumtable} \code{summ} option, and which works with \code{sumtable} \code{summ.names} styling.
}
\examples{
x <- 1:500
pctile(x)[50]
quantile(x,.5)
median(x)
}
