% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{propNA}
\alias{propNA}
\title{Proportion or number of missing values in a vector}
\usage{
propNA(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
This function calculates the proportion of values in a vector that are NA.
}
\details{
This function just shorthand for \code{mean(is.na(x))}, with a shorter name for reference in the \code{vtable} or \code{sumtable} \code{summ} option.
}
\examples{
x <- c(1, 1, NA, 2, 3, NA)
propNA(x)
}
