% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{countNA}
\alias{countNA}
\title{Number of missing values in a vector}
\usage{
countNA(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
This function calculates the number of values in a vector that are NA.
}
\details{
This function just shorthand for sum(is.na(x)), with a shorter name for reference in the vtable summ option.
}
\examples{
x <- c(1, 1, NA, 2, 3, NA)
countNA(x)
}
