% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nuniq}
\alias{nuniq}
\title{Number of unique values in a vector}
\usage{
nuniq(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
This function takes a vector and returns the number of unique values in that vector.
}
\details{
This function is just shorthand for \code{length(unique(x))}, with a shorter name for reference in the \code{vtable} or \code{sumtable} \code{summ} option.
}
\examples{
x <- c(1, 1, 2, 3, 4, 4, 4)
nuniq(x)

}
